Split rune funcs into separate file
